requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
numpy==2.0.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.13.0
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.5
sympy==1.13.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
